Fix imported target dependencies for OpenSSL
CMake V3.4 allows to search for the components SSL and Crypto
and defines targets for them.

The targets should be used instead of the old variables
as they populate the correct include path and libraries.

The targets should also be used in our interface library
definition so that it'll continue to work if OpenSSL was
moved after the cmake files was created.